fix: moved LandingPageHeading.tsx from suporter-plus-landing to digital-subscriber-checkout
