+++ support-frontend/assets/pages/digital-subscriber-checkout/digitalSubscriptionLandingPage.tsx
@@ -37,8 +37,8 @@ import {
 	UnitedStates,
 } from 'helpers/internationalisation/countryGroup';
 import { useContributionsSelector } from 'helpers/redux/storeHooks';
+import { LandingPageHeading } from 'pages/digital-subscriber-checkout/components/landingPageHeading';
 import { getPaymentMethodButtons } from 'pages/digital-subscriber-checkout/paymentButtons';
-import { LandingPageHeading } from 'pages/supporter-plus-landing/components/landingPageHeading';
 import { PaymentFailureMessage } from 'pages/supporter-plus-landing/components/paymentFailure';
 import { BillingPeriodSelector } from './components/billingPeriodSelector';
 import { PaymentTsAndCs } from './components/paymentTsAndCs';
