+++ support-frontend/assets/pages/digital-subscriber-checkout/preRenderDigitalSubscriptionLandingPage.tsx
@@ -24,7 +24,7 @@ import {
 	UnitedStates,
 } from 'helpers/internationalisation/countryGroup';
 import { PrerenderGlobalStyles } from 'helpers/rendering/prerenderGlobalStyles';
-import { LandingPageHeading } from 'pages/supporter-plus-landing/components/landingPageHeading';
+import { LandingPageHeading } from 'pages/digital-subscriber-checkout/components/landingPageHeading';
 
 const checkoutContainer = css`
 	position: relative;
